fix tag search
