+++ packages/web/src/components/search-bar/ConnectedSearchBar.jsx
@@ -107,21 +107,12 @@ class ConnectedSearchBar extends Component {
       }
     }
 
-    if (value.startsWith('#')) {
-      // perform tag search
-      this.props.history.push({
-        hash: value.split('#')[1],
-        pathname,
-        state: {}
-      })
-    } else {
-      value = encodeURIComponent(value)
-      this.props.history.push({
-        pathname: newPath,
-        search: locationSearchParams.toString(),
-        state: {}
-      })
-    }
+    value = encodeURIComponent(value)
+    this.props.history.push({
+      pathname: newPath,
+      search: locationSearchParams.toString(),
+      state: {}
+    })
   }
 
   onSelect = (value) => {
