+++ packages/web/src/pages/search-page-v2/types.ts
@@ -1,5 +1,19 @@
 import { IconComponent } from '@audius/harmony'
 
+export type ViewLayout = 'grid' | 'list'
+export const viewLayoutOptions: { label: string; value: ViewLayout }[] = [
+  { label: 'Grid', value: 'grid' },
+  { label: 'List', value: 'list' }
+]
+
+export enum CategoryView {
+  ALL = 'all',
+  PROFILES = 'profiles',
+  TRACKS = 'tracks',
+  PLAYLISTS = 'playlists',
+  ALBUMS = 'albums'
+}
+
 export type Filter =
   | 'genre'
   | 'mood'
